Cleanup

+++ src/commons/TargetTableEntry.h
@@ -2,7 +2,7 @@
 #define TARGET_TABLE_ENTRY_H
 struct __attribute__((__packed__)) TargetTableEntry
 {   
-    unsigned long kmerAsLong;
+    unsigned long long kmerAsLong;
     unsigned int sequenceID;
     unsigned int sequenceLength;
 };
